+++ tests/integration/test_urllib3.py
@@ -5,6 +5,7 @@
 import pytest
 import pytest_httpbin
 import vcr
+from vcr.patch import force_reset
 from assertions import assert_cassette_empty, assert_is_json
 urllib3 = pytest.importorskip("urllib3")
 
@@ -19,7 +20,7 @@ def verify_pool_mgr():
 
 @pytest.fixture(scope='module')
 def pool_mgr():
-    return urllib3.PoolManager()
+    return urllib3.PoolManager(cert_reqs='CERT_NONE')
 
 
 def test_status_code(httpbin_both, tmpdir, verify_pool_mgr):
@@ -55,7 +56,7 @@ def test_body(tmpdir, httpbin_both, verify_pool_mgr):
 def test_auth(tmpdir, httpbin_both, verify_pool_mgr):
     '''Ensure that we can handle basic auth'''
     auth = ('user', 'passwd')
-    headers = urllib3.util.make_headers(basic_auth='{0}:{1}'.format(*auth))
+    headers = urllib3.util.make_headers(basic_auth='{}:{}'.format(*auth))
     url = httpbin_both.url + '/basic-auth/user/passwd'
     with vcr.use_cassette(str(tmpdir.join('auth.yaml'))):
         one = verify_pool_mgr.request('GET', url, headers=headers)
@@ -69,7 +70,7 @@ def test_auth(tmpdir, httpbin_both, verify_pool_mgr):
 def test_auth_failed(tmpdir, httpbin_both, verify_pool_mgr):
     '''Ensure that we can save failed auth statuses'''
     auth = ('user', 'wrongwrongwrong')
-    headers = urllib3.util.make_headers(basic_auth='{0}:{1}'.format(*auth))
+    headers = urllib3.util.make_headers(basic_auth='{}:{}'.format(*auth))
     url = httpbin_both.url + '/basic-auth/user/passwd'
     with vcr.use_cassette(str(tmpdir.join('auth-failed.yaml'))) as cass:
         # Ensure that this is empty to begin with
@@ -138,3 +139,21 @@ def test_gzip(tmpdir, httpbin_both, verify_pool_mgr):
 def test_https_with_cert_validation_disabled(tmpdir, httpbin_secure, pool_mgr):
     with vcr.use_cassette(str(tmpdir.join('cert_validation_disabled.yaml'))):
         pool_mgr.request('GET', httpbin_secure.url)
+
+
+def test_urllib3_force_reset():
+    cpool = urllib3.connectionpool
+    http_original = cpool.HTTPConnection
+    https_original = cpool.HTTPSConnection
+    verified_https_original = cpool.VerifiedHTTPSConnection
+    with vcr.use_cassette(path='test'):
+        first_cassette_HTTPConnection = cpool.HTTPConnection
+        first_cassette_HTTPSConnection = cpool.HTTPSConnection
+        first_cassette_VerifiedHTTPSConnection = cpool.VerifiedHTTPSConnection
+        with force_reset():
+            assert cpool.HTTPConnection is http_original
+            assert cpool.HTTPSConnection is https_original
+            assert cpool.VerifiedHTTPSConnection is verified_https_original
+        assert cpool.HTTPConnection is first_cassette_HTTPConnection
+        assert cpool.HTTPSConnection is first_cassette_HTTPSConnection
+        assert cpool.VerifiedHTTPSConnection is first_cassette_VerifiedHTTPSConnection