+++ vcr/migration.py
@@ -59,7 +59,7 @@ def build_uri(**parts):
     port = parts['port']
     scheme = parts['protocol']
     default_port = {'https': 443, 'http': 80}[scheme]
-    parts['port'] = ':{0}'.format(port) if port != default_port else ''
+    parts['port'] = ':{}'.format(port) if port != default_port else ''
     return "{protocol}://{host}{port}{path}".format(**parts)
 
 
@@ -68,7 +68,7 @@ def _migrate(data):
     for item in data:
         req = item['request']
         res = item['response']
-        uri = dict((k, req.pop(k)) for k in PARTS)
+        uri = {k: req.pop(k) for k in PARTS}
         req['uri'] = build_uri(**uri)
         # convert headers to dict of lists
         headers = req['headers']
@@ -100,7 +100,7 @@ def migrate_json(in_fp, out_fp):
 
 
 def _list_of_tuples_to_dict(fs):
-    return dict((k, v) for k, v in fs[0])
+    return {k: v for k, v in fs[0]}
 
 
 def _already_migrated(data):
@@ -159,9 +159,9 @@ def main():
                  for (root, dirs, files) in os.walk(path)
                  for name in files)
     for file_path in files:
-            migrated = try_migrate(file_path)
-            status = 'OK' if migrated else 'FAIL'
-            sys.stderr.write("[{0}] {1}\n".format(status, file_path))
+        migrated = try_migrate(file_path)
+        status = 'OK' if migrated else 'FAIL'
+        sys.stderr.write("[{}] {}\n".format(status, file_path))
     sys.stderr.write("Done.\n")