+++ vcr/request.py
@@ -58,7 +58,10 @@ class Request(object):
         parse_uri = urlparse(self.uri)
         port = parse_uri.port
         if port is None:
-            port = {'https': 443, 'http': 80}[parse_uri.scheme]
+            try:
+                port = {'https': 443, 'http': 80}[parse_uri.scheme]
+            except KeyError:
+                pass
         return port
 
     @property
@@ -81,7 +84,7 @@ class Request(object):
         return self.scheme
 
     def __str__(self):
-        return "<Request ({0}) {1}>".format(self.method, self.uri)
+        return "<Request ({}) {}>".format(self.method, self.uri)
 
     def __repr__(self):
         return self.__str__()
@@ -91,7 +94,7 @@ class Request(object):
             'method': self.method,
             'uri': self.uri,
             'body': self.body,
-            'headers': dict(((k, [v]) for k, v in self.headers.items())),
+            'headers': {k: [v] for k, v in self.headers.items()},
         }
 
     @classmethod
@@ -112,7 +115,7 @@ class HeadersDict(CaseInsensitiveDict):
     In addition, some servers sometimes send the same header more than once,
     and httplib *can* deal with this situation.
 
-    Futhermore, I wanted to keep the request and response cassette format as
+    Furthermore, I wanted to keep the request and response cassette format as
     similar as possible.
 
     For this reason, in cassettes I keep a dict with lists as keys, but once